pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Denys-Kalinovskyi <ipzk211_kdyu@student.ztu.edu.ua>"]
readme = "README.md"
package-mode = false

[tool.poetry.scripts]
dev = "src.run:run"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "2.0.23"
python-dotenv = "1.0.0"
alembic = "1.13.0"
asyncpg = "0.29.0"
pydantic = "1.10.7"
fastapi = "0.110.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
starlette = "0.37.2"
requests = "2.31.0"
aiohttp = "3.9.2"
pytest = "7.4.0"
httpx = "0.24.1"
sqlalchemy-utils = "^0.41.2"
psycopg = "^3.1.18"
psycopg-binary = "^3.1.18"
boto3 = "^1.34.98"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
